feat: move common methods to src/pfsys/evm/mod.rs
