+++ src/lib.rs
@@ -1,6 +1,6 @@
 #![deny(
     bad_style,
-    dead_code,
+    // dead_code,
     improper_ctypes,
     non_shorthand_field_patterns,
     no_mangle_generic_items,
@@ -9,7 +9,6 @@
     patterns_in_fns_without_body,
     private_in_public,
     unconditional_recursion,
-    unused,
     unused_allocation,
     unused_comparisons,
     unused_parens,
@@ -28,7 +27,7 @@
     trivial_casts,
     trivial_numeric_casts,
     unused_extern_crates,
-    unused_import_braces,
+    // unused_import_braces,
     unused_qualifications
 )]
 #![feature(slice_flatten)]
